refactor: factor out AST
