+++ src/abi/AbiFunction.ts
@@ -1,4 +1,4 @@
-import { AstExpression } from "../grammar/ast";
+import { AstExpression } from "../ast/ast";
 import { CompilerContext } from "../context/context";
 import { WriterContext } from "../generator/Writer";
 import { TypeRef } from "../types/types";
