+++ src/generator/writers/writeExpression.spec.ts
@@ -4,11 +4,11 @@ import {
 } from "../../types/resolveDescriptors";
 import { WriterContext } from "../Writer";
 import { writeExpression } from "./writeExpression";
-import { openContext } from "../../grammar/store";
+import { openContext } from "../../context/store";
 import { resolveStatements } from "../../types/resolveStatements";
 import { CompilerContext } from "../../context/context";
 import { getParser } from "../../grammar";
-import { getAstFactory } from "../../grammar/ast";
+import { getAstFactory } from "../../ast/ast";
 import { defaultParser } from "../../grammar/grammar";
 
 const code = `
