+++ src/optimizer/standardOptimizer.ts
@@ -1,4 +1,4 @@
-import { AstExpression } from "../grammar/ast";
+import { AstExpression } from "../ast/ast";
 import {
     AddSelf,
     AddZero,
