+++ src/pipeline/build.ts
@@ -6,7 +6,7 @@ import { ConfigProject } from "../config/parseConfig";
 import { CompilerContext } from "../context/context";
 import { funcCompile } from "../func/funcCompile";
 import { writeReport } from "../generator/writeReport";
-import { getRawAST } from "../grammar/store";
+import { getRawAST } from "../context/store";
 import files from "../imports/stdlib";
 import { ILogger, Logger } from "../context/logger";
 import { PackageFileFormat } from "../packaging/fileFormat";
@@ -19,7 +19,7 @@ import { VirtualFileSystem } from "../vfs/VirtualFileSystem";
 import { compile } from "./compile";
 import { precompile } from "./precompile";
 import { getCompilerVersion } from "./version";
-import { FactoryAst, getAstFactory, idText } from "../grammar/ast";
+import { FactoryAst, getAstFactory, idText } from "../ast/ast";
 import { TactErrorCollection } from "../error/errors";
 import { getParser, Parser } from "../grammar";
 import { defaultParser } from "../grammar/grammar";
