+++ src/storage/resolveAllocation.spec.ts
@@ -1,13 +1,13 @@
 import fs from "fs";
 import { resolveDescriptors } from "../types/resolveDescriptors";
 import { getAllocations, resolveAllocations } from "./resolveAllocation";
-import { openContext } from "../grammar/store";
+import { openContext } from "../context/store";
 import { resolveStatements } from "../types/resolveStatements";
 import { CompilerContext } from "../context/context";
 import { resolveSignatures } from "../types/resolveSignatures";
 import path from "path";
 import { getParser } from "../grammar";
-import { getAstFactory } from "../grammar/ast";
+import { getAstFactory } from "../ast/ast";
 import { defaultParser } from "../grammar/grammar";
 
 const stdlibPath = path.resolve(__dirname, "../../stdlib/std/primitives.tact");
