+++ src/test/rename.spec.ts
@@ -1,11 +1,11 @@
 import fs from "fs";
 import { join } from "path";
-import { AstRenamer } from "../grammar/rename";
+import { AstRenamer } from "../ast/rename";
 import { prettyPrint } from "../prettyPrinter";
 import { trimTrailingCR, CONTRACTS_DIR } from "./util";
 import * as assert from "assert";
 import { getParser } from "../grammar";
-import { getAstFactory } from "../grammar/ast";
+import { getAstFactory } from "../ast/ast";
 import { defaultParser } from "../grammar/grammar";
 
 const EXPECTED_DIR = join(CONTRACTS_DIR, "renamer-expected");
