+++ src/types/resolveErrors.ts
@@ -1,7 +1,7 @@
 import { sha256_sync } from "@ton/crypto";
 import { CompilerContext, createContextStore } from "../context/context";
-import { AstNode, FactoryAst, isRequire } from "../grammar/ast";
-import { traverse } from "../grammar/iterators";
+import { AstNode, FactoryAst, isRequire } from "../ast/ast";
+import { traverse } from "../ast/iterators";
 import { evalConstantExpression } from "../optimizer/constEval";
 import { throwInternalCompilerError } from "../error/errors";
 import {
